Make solving even more robust in degenerate cases
Previously, when the centers don't describe a "full rank" matrix, then
the results aren't invariant to a transformation of the mean. This patch
always centers the mean at the origin. In particular, when there is just
one center, the result should be constant. There should be no effect
when the input is full rank (and the matrix has a true inverse).